Support device Stitching without host OpenCV
